requirements.txt
amqp==2.6.0
appdirs==1.4.3
asgiref==3.2.10
attrs==19.3.0
backcall==0.1.0
billiard==3.5.0.5
bleach==3.1.3
CacheControl==0.12.6
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.1
django-celery-beat==1.1.1
django-cors-headers==3.4.0
django-reset-migrations==0.3.1
django-taggit==1.2.0
djangorestframework==3.8.2
entrypoints==0.3
html5lib==1.0.1
idna==2.8
importlib-metadata==1.5.0
ipaddr==2.2.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
kombu==4.6.10
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
